+++ src/app/core/services/camera.service.ts
