+++ src/app/features/engine/service/material.service.ts
@@ -1,9 +1,9 @@
 import { inject, Injectable, signal } from '@angular/core';
-import vertexShader from '../shaders/vertexShader.glsl'
-import fragmentShader from '../shaders/fragmentShader.glsl'
+import vertexShader from '../shaders/slicerShader/vertexShader.glsl'
+import fragmentShader from '../shaders/slicerShader/fragmentShader.glsl'
 import * as THREE from 'three';
-import { SlicerService } from '../slicer/slicer.service';
-import { GeometryService } from '../geometry/geometry.service';
+import { SlicerService } from '../../slicer/service/slicer.service';
+import { GeometryService } from './geometry.service';
 
 @Injectable({
   providedIn: 'root'
@@ -17,6 +17,7 @@ export class MaterialService {
     this.material.set(new THREE.ShaderMaterial({
       vertexShader: vertexShader,
       fragmentShader: fragmentShader,
+      side: THREE.FrontSide,
       uniforms: {
         slicePlane: { value: slicePlane }
       }
