+++ src/app/features/engine/shaders/slicerShader/vertexShader.glsl
